feat(blackjackgame) : DIVIDE - BlackJackGame class
