+++ src/main/kotlin/blackjack/README.md
@@ -20,7 +20,7 @@
 - [X] 게임에 참여할 사람의 이름 입력 받기
 - [X] 배팅금액 입력 받기
 - [ ] 카드가 ACE 일 경우 예외처리 하기
-- [ ] 딜러와 참여자에게 카드 2장씩 나눠주기
+- [X] 딜러와 참여자에게 카드 2장씩 나눠주기
   - [ ] 받은 카드 노출하기
     - [ ] 카드 노출 시 딜러는 한 장만 보여주기
   - [ ] 처음 두 장의 카드 합이 21인 경우 배팅금액의 1.5배를 딜러에게 가져오기
