+++ src/test/kotlin/blackjack/BlackJackGameTest.kt
@@ -20,22 +20,3 @@ class BlackJackGameTest {
         assertThat(players.map { it.myCards.size }).isEqualTo(listOf(2, 2))
     }
 }
-
-class BlackJackGame(private val dealer: Gamer, private val players: List<Gamer>) {
-
-    fun initCardForDealer() {
-        repeat(FIRST_HAVE_NUMBER_OF_CARD) {
-            dealer.requestCard(Deck.pop())
-        }
-    }
-
-    fun initCardForPlayers() {
-        repeat(FIRST_HAVE_NUMBER_OF_CARD) {
-            players.map { it.requestCard(Deck.pop()) }
-        }
-    }
-
-    companion object {
-        private const val FIRST_HAVE_NUMBER_OF_CARD = 2
-    }
-}
